feat(nextbike/visualization): split math plotting in descriptive and predictive + add plotting method for classification
