+++ nextbike/visualization/__init__.py
@@ -1,2 +1,3 @@
-from .math import *
+from .math_descriptive import *
+from .math_predictive import *
 from .geo import *
